requirements.txt
asgiref==3.2.10
astroid==2.4.2
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.3
Django==3.1
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==8.1.0
pylint==2.5.3
python-decouple==3.3
pytz==2020.1
requests==2.25.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.26.2
wrapt==1.12.1
